Remove some duplication
