--- src/Idephix/File/FunctionBasedIdxFile.php.orig
@@ -1,69 +1,0 @@
-<?php
-namespace Idephix\File;
-
-use Idephix\File\Node\IdxTargetVisitor;
-use Idephix\IdxSetupCollector;
-use PhpParser\Lexer;
-use PhpParser\NodeTraverser;
-use PhpParser\Parser;
-
-class FunctionBasedIdxFile implements IdxFile
-{
-
-    /**
-     * @var Parser
-     */
-    private $parser;
-
-    public function __construct($file)
-    {
-        $this->setupCollector = new IdxSetupCollector();
-
-        $this->parser = new Parser(new Lexer());
-        $this->traverers = new NodeTraverser();
-<<<<<<< Updated upstream
-        $this->traverers->addVisitor(new IdxTargetVisitor($this->setupCollector));
-=======
-        $this->traverers->addVisitor(new IdxTargetVisitors($this->setupCollector));
-        $this->traverers->addVisitor(new IdxTaskVisitor($this->setupCollector));
->>>>>>> Stashed changes
-
-        $stmts = $this->parser->parse(file_get_contents($file));
-        $this->traverers->traverse($stmts);
-    }
-
-    public function targets()
-    {
-        return $this->setupCollector->getTargets();
-    }
-
-    public function sshClient()
-    {
-        // TODO: Implement sshClient() method.
-    }
-
-    public function output()
-    {
-        // TODO: Implement output() method.
-    }
-
-    public function input()
-    {
-        // TODO: Implement input() method.
-    }
-
-    public function tasks()
-    {
-        return [
-            'foo' => function(){
-                echo 'bar';
-            }
-        ];
-    }
-
-    public function libraries()
-    {
-        // TODO: Implement libraries() method.
-    }
-
-}
